==> analysis/data-processing/02_data-processing.Rmd <==
---
title: "Data processing and analysis data generation"
author: "Bryan Mayer"
date: "`r Sys.Date()`"
output: html_document
---

```{r load-packages, echo = FALSE, message=F,output=F}
knitr::opts_chunk$set(echo = T)
library(VISCfunctions)
library(janitor)
library(lubridate)
library(tidyverse)
library(here)
library(glue)
library(kableExtra)
library(patchwork)

theme_set(theme_bw() + theme(legend.position = "top"))

options(scipen = 999)

# plot cutoff
cutoff = 1.602060
cutoffexp = substitute(expression(NULL <= a), list(a = cutoff))

sum_nonmissing = function(x) sum(!is.na(x))
diffdays = function(x, y) as.numeric(difftime(mdy(x), mdy(y), units = "days"))


rx_lab = tibble(
  rx_code2 = c("C", "T1", "T2"),
  trt_cat = c("Placebo", "VRC01 (10mg/kg)", "VRC01 (30mg/kg)")
)

shared_drive = dir.exists("/Volumes/trials/")

save_shared_folder = F
if(!save_shared_folder) print("DATA WAS NOT UPDATED ON SHARED FOLDER")

if(!dir.exists(here("output/data-processing"))) dir.create(here("output/data-processing"))

```


```{r load-funs-data}
source(here("R", "directory-funs.R"))
```

```{r load-data}

neut_data = read_csv(clean_data_here("amp-neut-blinded.csv"), col_types = cols())
vl_data_all = read_csv(clean_data_here("full-vl-data.csv"), col_types = cols())
time_data = read_csv(clean_data_here("ptid-time-data.csv"), col_types = cols())
pk_data_nm = read_csv(clean_data_here("pk-nm-data.csv"), col_types = cols())
raw_dose_data = read_csv(raw_data_here("pk-dosing-info.csv"))
placebo_dose_data = read_csv(raw_data_here("placebo-dosing-all.csv"))
rx_dat = read_csv(raw_data_here("rx_dat.csv"), col_types = cols())
infected_ids = rx_dat %>% dplyr::filter(pub_id %in% neut_data$pub_id) %>% pull(pub_id)

```


## Overview and roadmap

There are four pieces to this processing:

1) The neutralization data
2) The viral load data
    - determine final filtered set
    -  unblind
3) The PK data
    - de-monolix
    - unblind
    - restrict to filtered
    - merge on additional relevant data (time/demo)
    - create time variables
4) Time summary data
5) final datasets - merge on the neut to PK and VL data, save


The goal of this script is to QC the cleaned version of those datasets to determine the final analysis data with the final cohort of participants. Specific

For NLME modeling, additional processing is required and done in a subsequent markdown (`03_create-mlx-data.Rmd`). 

## Background

From the NEJM paper:

Table 2: There were 67 placebo infections and 107 VRC01 infections (174 total ptids)
Figure 2: There were 64 placebo IC80s (19+10+35) and 98 VRC01 IC80s (70 + 19 + 9) (162 total ptids)

## Notes from raw data processing

## Neutralization data

For neutralization, this data was already gently cleaned. The raw neutralization data contained 174 ptids, but 12 of those were designated nisolates = 0 (missing gmt80 measurements) and were excluded. The remaining 162 total participants matches NEJM sample size.

No additional processing was needed for the neut data.

```{r neut-exploration}

glimpse(neut_data)

neut_data %>%
  dplyr::select(-pub_id, -nisolates) %>%
  summarize(across(everything(), list(n = sum_nonmissing)))

```

```{r neut-by-trt}

neut_data %>% 
  left_join(rx_dat, by = "pub_id") %>%
  group_by(trt = rx_code2) %>%
  summarize(n(), .groups = "drop") %>%
  kable(caption="Neut. data sample size by trt") %>% kable_styling(full_width = F)

```

## Viral load data

```{r vl-exploration}

glimpse(vl_data_all)

```

### Sample size assessment

There are 220 participants in this dataset: so 46 not included in the infection list used in the NEJM paper. These participants were all infected at least 86 weeks following enrollment and didn't meet primary endpoint requirements. The 12 ptids in the neut data with missing data are in  the viral load dataset but will be removed because they lack the neutralization data.

The `has_neut_data` and `infection_pre85wks` flags were previously merged onto these data and confirmed that 162 participants remain when the data is filtered on these two flags.

```{r vl-unblind}

vl_data_unblind = vl_data_all %>% 
  left_join(rx_dat, by = "pub_id")

```

There's one case of a participant with a negative enrollment day (-1). They have no PK data, so just going to augment to 
days_enrollment = 0.

```{r vl-neg-enrollment}

nrow(subset(vl_data_unblind, days_enrollment < 0))
subset(vl_data_unblind, pub_id == "703-0795") %>% select(pub_id, days_enrollment, fp_day, dx_day, vl)
subset(pk_data_nm, pub_id == "703-0795")

x = vl_data_unblind$days_enrollment
vl_data_unblind = mutate(vl_data_unblind, days_enrollment = if_else(pub_id == "703-0795", 
                                                  pmax(days_enrollment, 0), days_enrollment))
stopifnot(sum(x != vl_data_unblind$days_enrollment) == 1 & 
            setdiff(x, vl_data_unblind$days_enrollment) == -1)


```

```{r vl-full-sample}

vl_data_unblind %>% distinct(pub_id, infection_pre85wks, has_neut_data) %>%
  group_by(infection_pre85wks, has_neut_data) %>% 
  summarize(n(), .groups = "drop") %>%
  kable() %>% kable_styling(full_width = F)

```


**Data will be additionally restricted to ptids who have at least 2 detectable VL measurements starting with first positive and ending with  ART.**

Among the 162 participants, 4 do not have enough data for analysis leaving 158 ptids. Note: a previous version of this analysis used diagnosis time, but this was later than first positive for 31 participants.

```{r}

measurement_summary = vl_data_unblind %>%
  filter(days_fp >= 0 & infection_pre85wks & has_neut_data) %>%
   group_by(pub_id, has_neut_data, protocol) %>%
  summarize(
    peak_vl = max(vl),
    peak_vl_no_art = max(vl[on_art == 0]),
    total_measurements = n(),
    total_measurements_offart = length(vl[on_art == 0]),
    total_measurements_pos = length(vl[cens != 1]),
    total_pos_offart = length(vl[on_art == 0 & cens != 1]),
    total_pos_onart = n() - total_measurements_pos
  , .groups = "drop") %>%
  gather(key, total, -pub_id, -peak_vl, -peak_vl_no_art, -has_neut_data, -protocol) 

with(measurement_summary, ftable(key))

measurement_summary %>%
  dplyr::filter(key == "total_pos_offart") %>%
  mutate(total_pos_vl = pmin(total, 5)) %>%
  group_by(key, total_pos_vl) %>%
  summarize(
    total_ptids = n()
    , .groups = "drop") %>%
  mutate(total_pos_vl = if_else(total_pos_vl < 5, as.character(total_pos_vl), "> 4")) %>%
  select(-key) %>%
  adorn_totals() %>%
  write_csv(here("output/data-processing", "measurement-sample-size.csv")) %>%
  kable(caption = "Sample sizes for total detectable VL (no ART) - Table saved in output/data-processing") %>% kable_styling(full_width = F)

measurement_summary %>%
  ggplot(aes(x = total, fill = factor(has_neut_data))) +
  geom_histogram(binwidth = 1) +
  facet_wrap(~key) +
  scale_x_continuous("VL measurements per ptid (post dx)", breaks = 0:15) +
  scale_y_continuous("Total ptids", breaks = 10 * 0:15)

measurement_summary %>%
  dplyr::filter(key == "total_pos_offart") %>%
  summarize(
    total_ptids = n(),
    total_excluded = sum(total < 2),
    total_included = sum(total >= 2)
    , .groups = "drop") %>%
  kable(caption = "Sample sizes (ppts with neut data) for analysis (2+ uncensored VLs, no ART)") %>% kable_styling(full_width = F)

measurement_summary %>%
  dplyr::filter(key == "total_pos_offart") %>%
  gather(type, peak_vl, peak_vl:peak_vl_no_art) %>%
  ggplot(aes(x = type, y = peak_vl, colour = type)) +
  geom_boxplot() +
  geom_point() +geom_line(aes(group = pub_id)) +
  facet_wrap(~total, nrow = 1, strip.position = "bottom") +
  scale_x_discrete("Off-ART VL measurements per ptid (post dx)", breaks = 0:15) +
  scale_y_log10("Peak VL (any measurement)", breaks = 10^(2:7)) +
  theme(strip.placement = "outside", strip.background = element_blank(), panel.spacing = unit(0, "lines"),
        panel.border = element_rect(colour = "gray"))

```

```{r vl-data-trt}

vl_data_unblind %>%
  distinct(pub_id, infection_pre85wks, has_neut_data, trt = rx_code2) %>%
  group_by(infection_pre85wks, has_neut_data, trt) %>% 
  summarize(n(), .groups = "drop") %>%
  pivot_wider(names_from = "trt", values_from = "n()") %>%
  kable() %>% kable_styling(full_width = F)

measurement_summary %>%
  dplyr::filter(key == "total_pos_offart") %>%
  left_join(rx_dat, by = "pub_id") %>%
  mutate(total_pos_vl = pmin(total, 5)) %>%
  group_by(key, total_pos_vl, trt = rx_code2) %>%
  summarize(
    total_ptids = n()
    , .groups = "drop") %>%
  pivot_wider(names_from = "trt", values_from = "total_ptids") %>%
  mutate(total_pos_vl = if_else(total_pos_vl < 5, as.character(total_pos_vl), "> 4")) %>%
  select(-key) %>%
  adorn_totals() %>%
  write_csv(here("output/data-processing", "measurement-sample-size_grp.csv")) %>%
  kable(caption = "Sample sizes for total detectable VL (no ART) - saved to output/data-processing") %>% kable_styling(full_width = F)

measurement_summary %>%
  dplyr::filter(key == "total_pos_offart") %>%
  left_join(rx_dat, by = "pub_id") %>%
  group_by(trt = rx_code2) %>%
  summarize(
    total_ptids = n(),
    total_excluded = sum(total < 2),
    total_included = sum(total >= 2)
    , .groups = "drop") %>%
  kable(caption = "Sample sizes (ppts with neut data) for analysis (2+ uncensored VLs, no ART)") %>% kable_styling(full_width = F)


```

```{r analysis-pubids}

low_sample_ptids = measurement_summary %>%
  dplyr::filter(key == "total_pos_offart" & total <= 1) %>%
  select(pub_id) %>%
  unlist() %>% unname()

strict_sample_ptids = measurement_summary %>%
  dplyr::filter(key == "total_pos_offart" & total <= 2) %>%
  select(pub_id) %>%
  unlist() %>% unname()

```


Neut. data for the low measurement sample participants.

```{r vl-neut-exploration}

neut_data %>%
  subset(pub_id %in% low_sample_ptids) %>%
  select(pub_id, nisolates, gmt80mf)

```

## PK Data

### Dose data

This contains all infusion dates, rates, and doses. Merge on the diagnosis time, add a time variable and subset to analysis cohort.

There are many participants who received an infusion on their diagnosis date (52). There are 9 rows post-fp but pre-dx.

```{r dosing-data}

dose_data = raw_dose_data %>%
  mutate(has_PK = T) %>%
  bind_rows(mutate(placebo_dose_data, has_PK = F)) %>%
  dplyr::filter(pub_id %in% neut_data$pub_id) %>%
  left_join(select(time_data, pub_id, enrdt, dxdt, fpdt), by = "pub_id") %>%
  mutate(infusiondt = as.character(format(infusiondt, "%m-%d-%Y"))) %>%
  group_by(pub_id) %>%
  mutate(
    days_enrollment = diffdays(infusiondt, enrdt),
    days_enrollment_dose = if_else(days_enrollment == 0, days_enrollment, days_enrollment + 0.1),
    days_fp =  diffdays(infusiondt, fpdt),
    days_dx =  diffdays(infusiondt, dxdt),
    post_dx = days_dx >= 0,
    post_fp = days_fp >= 0,
    final_infusiondt = format(max(mdy(infusiondt)), "%m-%d-%Y"),
    final_infusiondt_pre_dx = format(max(mdy(infusiondt[!post_dx])), "%m-%d-%Y"),
    final_infusiondt_pre_fp = format(max(mdy(infusiondt[!post_fp])), "%m-%d-%Y"),
    final_infusion_pre_dx = infusiondt == as.character(final_infusiondt_pre_dx),
    final_infusion_pre_fp = infusiondt == as.character(final_infusiondt_pre_fp)
  ) %>%
  ungroup()

stopifnot(n_distinct(subset(dose_data, has_PK)$pub_id) == 98)
stopifnot(n_distinct(dose_data$pub_id) == 162)

dose_data %>%
  group_by(has_PK) %>%
  summarize(
    sum(dose_data$days_dx == 0),
    sum(dose_data$days_dx > 0),
    n_distinct(pub_id[days_dx >= 0])
  )

dose_data %>%
  group_by(has_PK) %>%
  summarize(
    sum(dose_data$days_fp == 0),
    sum(dose_data$days_fp > 0),
    n_distinct(pub_id[days_fp >= 0])
  )

stopifnot(nrow(subset(dose_data, final_infusion_pre_dx)) == 162)
stopifnot(n_distinct(subset(dose_data, final_infusion_pre_dx)$pub_id) == 162)

stopifnot(nrow(subset(dose_data, final_infusion_pre_fp)) == 162)
stopifnot(n_distinct(subset(dose_data, final_infusion_pre_fp)$pub_id) == 162)

```

### Concentration data

This data has been carefully processed by the hvtn. It is case-control data (1:1), so there will be uninfected but treated participants in these data. However, there are exclusionary rules for the case-control data that are out of sync with the actual infusion schedule of the participants through acquisition.

As expected, there are data for the 98 ptids with available neutralization data matching the 98 treated ptids in the VL data.

```{r nm-pk-overview}

glimpse(pk_data_nm)

pk_data_nm = pk_data_nm %>%
  mutate(
    infection_pre85wks = pub_id %in% unique(subset(vl_data_all, infection_pre85wks & has_neut_data)$pub_id),
    has_neut_data = pub_id %in% neut_data$pub_id
    )

pk_data_nm %>% distinct(pub_id, infection_pre85wks, has_neut_data) %>%
  group_by(infection_pre85wks, has_neut_data) %>% 
  summarize(n(), .groups = "drop") %>%
  kable() %>% kable_styling(full_width = F)

pk_data_nm %>%
  mutate(tmp = pub_id %in% neut_data$pub_id) %>%
  dplyr::filter(tmp) %>%
  pull(pub_id) %>%
  n_distinct() 

n_distinct(subset(pk_data_nm, has_neut_data)$pub_id)

```

To create the analysis dataset (and eventual PKPD dataset), the processing must:

 - de-nonmem/monolix the data: remove the '.''s, collapse the staggered dosing variables, drop ID, rename the variables
 - unblind (it's already technically unblinded), but can verify the unblinding matches
 - create the relevant time variables
 - restrict to ptids with neut/vl data filter out those with limited viral load data (should just be two ptids): 96 ptids

```{r pk-data-clean}

 pk_data_stem = pk_data_nm %>%
  dplyr::filter(has_neut_data & infection_pre85wks) %>%
  # remove . = NA, convert back to numeric
  mutate(across(everything(), ~na_if(., "."))) %>%
  mutate(across(c(AMT, DV, RATE), as.numeric)) %>%
  mutate(
    dose_row = !is.na(AMT) & is.na(DV) #remove these
    ) %>%
  group_by(ID) %>%
  mutate(
    no_followup = all(dose_row), # 15 people drop out after first infusion
    final_dose_row = dose_row & TIME == max(TIME) # want to keep these rows, sort of like no-followup
    ) %>% 
  # collapse dosing rows into long-format
  group_by(ID, DOSENO) %>%
  arrange(-dose_row, TIME) %>% # fixes time issues
  mutate(
    dose_mg = zoo::na.locf(AMT),
    infustion_rate = zoo::na.locf(RATE),
    infusion_day = min(TIME) 
  ) %>%
  ungroup() %>%
  arrange(ID, TIME) %>%
  # remove dose rows, rename and remove variables
  dplyr::filter(!dose_row | no_followup | final_dose_row) %>%
  dplyr::select(pub_id, everything(), -ID, -AMT, -AVISITN, -RATE, -dose_row, -final_dose_row) %>%
  rename(days_enrollment = TIME, concentration = DV, infusion_no = DOSENO) %>%
  left_join(rx_dat, by = "pub_id") %>%
  left_join(dplyr::select(time_data, -has_neut_data, -infection_pre85wks), by = "pub_id") %>%
  mutate(
    days_fp = days_enrollment - fp_day,
    days_dx = days_enrollment - dx_day,
    days_tpf = days_enrollment - tpf_day
  )

glimpse(pk_data_stem)

n_distinct(pk_data_stem$pub_id)
n_distinct(subset(pk_data_stem, no_followup)$pub_id)
n_distinct(subset(pk_data_stem, !is.na(concentration))$pub_id)

```

Time-diagnostics:

```{r}
range(pk_data_stem$days_dx)
range(pk_data_stem$days_fp)
range(pk_data_stem$days_tpf)
```

Other variables:

```{r pk-data-checks}

with(pk_data_stem, ftable(study, protocol))
with(pk_data_stem, ftable(dose, rx_code2)) # can drop this variable

```

## Time Summary

The `time_data` dataset was created upstream. Here, we subset out based on has_neut_data & infection_pre85wks

Then, we add some key variables:

 - final infusion day from clean dose data (above)
 - final_vl_day is max(enrollment_day) based on the vl dataset, rename as final_day [x]
 - from the cleaned vl data:
   - regional cat [x]
   - final negative day (before dx) [x]
   - rx info, then create rx lab variable [x]

There are 11 cases of mismatch with the PK data, where the final pre-dx infusion is later.
9/11 of those cases correspond to first positive exclusionary criteria (fp  before dx retrospective). 
The other two cases, not sure, maybe a big delay for dx.

```{r get-final-infusion-info}

time_data_test = left_join(time_data, rx_dat, by = "pub_id")
mismatch_time_ptids = dplyr::filter(time_data_test, dx_day != fp_day)

final_infusion_data = dose_data %>%
  group_by(pub_id, has_PK) %>%
  summarize(
    final_infusion_day = days_enrollment[which.max(infusion_no)],
    final_infusion_pre_dx_day = days_enrollment[final_infusion_pre_dx],
    final_infusion_pre_fp_day = days_enrollment[final_infusion_pre_fp],
    total_infusions = max(infusion_no),
    total_infusions_pre_dx = max(infusion_no[!post_dx]),
    total_infusions_pre_fp = max(infusion_no[!post_fp]),
    .groups = "drop"
  ) 


stopifnot(nrow(final_infusion_data)== 162)
stopifnot(n_distinct(final_infusion_data$pub_id)== 162)

stopifnot(all(final_infusion_data$final_infusion_day >= final_infusion_data$final_infusion_pre_dx_day))
stopifnot(all(final_infusion_data$final_infusion_pre_dx_day >= final_infusion_data$final_infusion_pre_fp_day))

stopifnot(all(final_infusion_data$total_infusions >= final_infusion_data$total_infusions_pre_dx))
stopifnot(all(final_infusion_data$total_infusions_pre_dx >= final_infusion_data$total_infusions_pre_fp))
stopifnot(all(final_infusion_data$total_infusions_pre_fp >= 1))

pk_data_final_infusions = pk_data_stem %>%
  group_by(pub_id) %>%
  summarize(
    final_infusion_day = round(max(days_enrollment)),
    total_infusions = max(infusion_no),
    .groups = "drop"
  ) 

final_infusion_data %>%
  mutate(has_PK) %>%
  dplyr::select(pub_id, final_infusion_pre_dx_day, total_infusions_pre_dx, final_infusion_pre_fp_day) %>%
  full_join(pk_data_final_infusions, by = "pub_id") %>%
  dplyr::filter(total_infusions_pre_dx != total_infusions)

final_infusion_data %>%
  dplyr::filter(pub_id %in% c("704-2624", "704-2901"))
      
```



```{r vl-time-summary}

# lose two cases, the two without viral loads
vl_time_summary = vl_data_unblind %>%
  dplyr::filter(days_fp < 0) %>%
  group_by(pub_id, study_region_cat, dx_day, fp_day, rx_code2) %>%
  summarize(
    last_neg_day = max(days_enrollment),
    .groups = "drop"
  )

# the two without fp data, no viral loads
stopifnot(!unique(subset(vl_data_unblind, is.na(fp_day))$pub_id) %in% vl_time_summary$pub_id)
  
range(vl_time_summary$last_neg_day)
stopifnot(all(vl_time_summary$dx_day > vl_time_summary$last_neg_day))
stopifnot(all(vl_time_summary$fp_day > vl_time_summary$last_neg_day))

```

```{r create-time-summary}

time_summary_full = time_data %>%
  rename(final_day = final_vl_day) %>%
  left_join(dplyr::select(vl_time_summary, -dx_day, -fp_day), by = "pub_id") %>%
  left_join(select(final_infusion_data, -contains("total_"), -has_PK), by = "pub_id")

```

## Create final analysis datasets

```{r}

neut_merge = dplyr::select(neut_data, -contains("cat")) 
time_merge = time_summary_full %>%
  mutate(
    final_preart_day = pmin(art_day, final_day),
    post_fp_infusion = final_infusion_day >= fp_day
  ) %>%
  dplyr::select(pub_id, final_infusion_day, final_day, final_preart_day, post_fp_infusion)

vl_data_analysis_full = vl_data_unblind %>%
  dplyr::filter(on_art == 0 & has_neut_data & infection_pre85wks) %>%
  left_join(neut_merge, by = "pub_id") %>%
  left_join(time_merge, by = "pub_id",  suffix = c("_timeerrorX", "_timeerrorY")) %>%
  mutate(log10vl = log10(vl), 
        vl_model_flag = !pub_id %in% low_sample_ptids, 
        rx_lab = factor(rx_code2, levels = rx_lab$rx_code2, labels = rx_lab$trt_cat))  

n_distinct(vl_data_analysis_full$pub_id)
stopifnot(n_distinct(vl_data_analysis_full$pub_id) == 162)
stopifnot(!str_detect(names(vl_data_analysis_full), "_timeerrorX")) #duplicate merge catch

stats_ptids = unique(vl_data_analysis_full$pub_id) 

vl_data_analysis = vl_data_analysis_full %>%
  dplyr::filter(vl_model_flag)

n_distinct(vl_data_analysis$pub_id)
stopifnot(n_distinct(vl_data_analysis$pub_id) == 158)

pk_data_analysis = pk_data_stem %>% 
  dplyr::filter(has_neut_data & infection_pre85wks) %>%
  left_join(neut_merge, by = "pub_id") %>%
  mutate(
    stats_flag = pub_id %in% stats_ptids,
    vl_model_flag = !pub_id %in% low_sample_ptids, 
    rx_lab = factor(rx_code2, levels = rx_lab$rx_code2, labels = rx_lab$trt_cat)
    )

n_distinct(pk_data_analysis$pub_id)
stopifnot(n_distinct(pk_data_analysis$pub_id) == 98)
n_distinct(subset(pk_data_analysis, stats_flag)$pub_id)
n_distinct(subset(pk_data_analysis, vl_model_flag)$pub_id)

time_analysis = time_summary_full %>%
  dplyr::filter(has_neut_data & infection_pre85wks) %>%
  left_join(neut_merge, by = "pub_id") %>%
  mutate(
    stats_flag = pub_id %in% stats_ptids,
    rx_lab = factor(rx_code2, levels = rx_lab$rx_code2, labels = rx_lab$trt_cat),
    vl_model_flag = !pub_id %in% low_sample_ptids
    )

n_distinct(time_analysis$pub_id)
stopifnot(n_distinct(time_analysis$pub_id) == 162)
stopifnot(all(!is.na(time_analysis$final_infusion_day)))
stopifnot(n_distinct(subset(time_analysis, stats_flag)$pub_id) == 162)
stopifnot(n_distinct(subset(time_analysis, vl_model_flag)$pub_id) == 158)

```


```{r save-final-datasets}

measurement_summary %>%
  glimpse() %>%
  write_csv(clean_data_here("ptid-measurement-counts.csv"))

dplyr::select(vl_data_analysis_full, -flag) %>%
  glimpse() %>%
  write_csv(clean_data_here("adata-vl-stats.csv"))

dplyr::select(vl_data_analysis, -flag) %>%
  glimpse() %>%
  write_csv(clean_data_here("adata-vl.csv"))

stopifnot(n_distinct(subset(dose_data, has_PK)$pub_id) == 98)
dose_data %>%
  dplyr::filter(has_PK) %>%
  dplyr::select(-has_PK) %>%
  glimpse() %>%
  write_csv(clean_data_here("dosing-data.csv"))

dplyr::select(pk_data_analysis, -dose) %>%
  glimpse() %>%
  write_csv(clean_data_here("adata-pk.csv"))

time_analysis %>%
  glimpse() %>%
  write_csv(clean_data_here("adata-time-summary.csv"))

```


```{r save-shared, eval = save_shared_folder}

write_csv(tibble(), 
           paste0("/Volumes/trials/vaccine/Bryan_Mayer/AMP-Modeling/data/!DATA UPDATED - ",
                  Sys.Date()))

write_csv(dplyr::select(vl_data_analysis, -flag, -has_neut_data, -infection_pre85wks, -rx),
                         "/Volumes/trials/vaccine/Bryan_Mayer/AMP-Modeling/data/adata-viral-loads.csv")

write_csv(dplyr::select(pk_data_analysis, -dose, -has_neut_data, -infection_pre85wks, -rx),
                         "/Volumes/trials/vaccine/Bryan_Mayer/AMP-Modeling/data/adata-pk.csv")

write_csv(dplyr::select(time_analysis, -has_neut_data, -infection_pre85wks),
                         "/Volumes/trials/vaccine/Bryan_Mayer/AMP-Modeling/data/adata-time-summary.csv")

```
